move ConcreteBuffer implementation to CPP file
