+++ cpp/modmesh/buffer/CMakeLists.txt
@@ -15,6 +15,7 @@ set(MODMESH_BUFFER_HEADERS
 
 set(MODMESH_BUFFER_SOURCES
     ${CMAKE_CURRENT_SOURCE_DIR}/BufferExpander.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/ConcreteBuffer.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/SimpleArray.cpp
     CACHE FILEPATH "" FORCE)
